% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{sew}
\alias{sew}
\title{Estimation of the sensitivity with complex survey data}
\usage{
sew(response.var, phat.var, weights.var, event, cutoff.point, data = NULL)
}
\arguments{
\item{response.var}{vector with information of the response variable
or character string indicating the name of the variable
indicating the response variable in the data set}

\item{phat.var}{vector with information of the estimated probabilities
or character string indicating the name of the variable
indicating these probabilities in the data set}

\item{weights.var}{vector with information of the sampling weights
or character string indicating the name of the variable
indicating the sampling weights in the data set}

\item{event}{label indicating the event of interest (default \code{event = 1})}

\item{cutoff.point}{a numeric value indicating the cut-off point to be used}

\item{data}{name of the data set.
If data=NULL, then specific numerical vectors must be included in
\code{response.var}, \code{phat.var} and \code{weights.var}}
}
\value{
A numeric value indicating the estimated sensitivity parameter considering sampling weights
}
\description{
Estimate the sensitivity parameter
considering sampling weights with complex survey data
}
\examples{
data(example_data_wroc)

sew(response.var = "y", phat.var = "phat", weights.var = "weights",
    event = 1, cutoff.point = 0.5, data = example_data_wroc)

# Or equivalently
sew(response.var = example_data_wroc$y,
    phat.var = example_data_wroc$phat,
    weights.var = example_data_wroc$weights,
    event = 1, cutoff.point = 0.5)
}
