% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wocp.R
\name{wocp}
\alias{wocp}
\title{Optimal cut-off points for complex survey data}
\usage{
wocp(
  response.var,
  phat.var,
  weights.var = NULL,
  tag.event = NULL,
  tag.nonevent = NULL,
  method = c("Youden", "MaxProdSpSe", "ROC01", "MaxEfficiency"),
  data = NULL,
  design = NULL
)
}
\arguments{
\item{response.var}{Vector with information of the response variable
or character string with the name of the variable
indicating the response variable in the data set.}

\item{phat.var}{Vector with information of the estimated probabilities
or character string with the name of the variable
indicating these probabilities in the data set.}

\item{weights.var}{Vector with information of the sampling weights
or character string with the name of the variable
indicating the sampling weights in the data set.
If \code{NULL}, the sampling design should be specified in the argument \code{design}.
For unweighted estimates, set all the sampling weight values to 1.}

\item{tag.event}{Label indicating the event of interest in \code{response.var}.
The default option is \code{tag.event = NULL}, which selects the class with the lowest number of units.}

\item{tag.nonevent}{Label indicating the non-event in \code{response.var}.
The default option is \code{tag.nonevent = NULL}, which selects the class with the greatest number of units.}

\item{method}{Choose one of the following methods (Lopez-Raton et al, 2014) to obtain the optimal cut-off point:
\code{MaxProdSpSe}, \code{ROC01}, \code{Youden}, \code{MaxEfficiency}}

\item{data}{Data frame which must incorporate information on the columns
\code{response.var}, \code{phat.var} and \code{weights.var}.
If \code{data=NULL}, then specific numerical vectors must be included in
\code{response.var}, \code{phat.var} and \code{weights.var}
or the sampling design should be indicated in the argument \code{design}.}

\item{design}{An object of class \code{survey.design} generated by
\code{survey::svydesign} indicating the complex sampling design of the data.
If \code{NULL} information on the data and weights must be included in the argument
\code{data} or as a vector in the argument \code{weights.var}.}
}
\value{
an object of class \code{wocp}. This object is a list that contains information about the method used to calculate the optimal cut-off point, the optimal value of the cut-off point and the corresponding value of the selected optimality criterion, as well as all the cut-off points considered and the corresponding values of the criterion.
}
\description{
Calculate optimal cut-off points for complex survey data (Iparragirre et al., 2022).
Some functions of the package OptimalCutpoints (Lopez-Raton et al, 2014) have been used and modified in order them to consider sampling weights.
}
\examples{
data(example_data_wroc)

myocp <- wocp(response.var = "y", phat.var = "phat", weights.var = "weights",
              tag.event = 1, tag.nonevent = 0,
              method = "Youden",
              data = example_data_wroc)

# Or equivalently
myocp <- wocp(example_data_wroc$y, example_data_wroc$phat, example_data_wroc$weights,
              tag.event = 1, tag.nonevent = 0, method = "Youden")


}
\references{
Iparragirre, A., Barrio, I., Aramendi, J. and Arostegui, I. (2022). Estimation of cut-off points under complex-sampling design data. \emph{SORT-Statistics and Operations Research Transactions} \bold{46}(1), 137--158.

Lopez-Raton, M., Rodriguez-Alvarez, M.X, Cadarso-Suarez, C. and Gude-Sampedro, F. (2014). OptimalCutpoints: An R Package for Selecting Optimal Cutpoints in Diagnostic Tests. \emph{Journal of Statistical Software} \bold{61}(8), 1--36.
}
