% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsp.R
\name{wsp}
\alias{wsp}
\title{Estimation of the specificity with complex survey data}
\usage{
wsp(
  response.var,
  phat.var,
  weights.var = NULL,
  tag.nonevent = NULL,
  cutoff.value,
  data = NULL,
  design = NULL
)
}
\arguments{
\item{response.var}{Vector with information of the response variable
or character string with the name of the variable
indicating the response variable in the data set.}

\item{phat.var}{Vector with information of the estimated probabilities
or character string with the name of the variable
indicating these probabilities in the data set.}

\item{weights.var}{Vector with information of the sampling weights
or character string with the name of the variable
indicating the sampling weights in the data set.
If \code{NULL}, the sampling design should be specified in the argument \code{design}.
For unweighted estimates, set all the sampling weight values to 1.}

\item{tag.nonevent}{Label indicating the non-event in \code{response.var}.
The default option is \code{tag.nonevent = NULL}, which selects the class with the greatest number of units.}

\item{cutoff.value}{A numeric value indicating the cut-off point to be used.
No default value for this argument. A numeric value must be indicated necessarily.}

\item{data}{Data frame which must incorporate information on the columns
\code{response.var}, \code{phat.var} and \code{weights.var}.
If \code{data=NULL}, then specific numerical vectors must be included in
\code{response.var}, \code{phat.var} and \code{weights.var}
or the sampling design should be indicated in the argument \code{design}.}

\item{design}{An object of class \code{survey.design} generated by
\code{survey::svydesign} indicating the complex sampling design of the data.
If \code{NULL} information on the data and weights must be included in the argument
\code{data} or as a vector in the argument \code{weights.var}.}
}
\value{
A numeric value indicating the estimated specificity parameter considering sampling weights
}
\description{
Estimate the specificity parameter for a given cut-off point
considering sampling weights with complex survey data.
}
\examples{
data(example_data_wroc)

sp.obj <- wsp(response.var = "y",
              phat.var = "phat",
              weights.var = "weights",
              tag.nonevent = 0,
              cutoff.value = 0.5,
              data = example_data_wroc)

# Or equivalently
sp.obj <- wsp(response.var = example_data_wroc$y,
              phat.var = example_data_wroc$phat,
              weights.var = example_data_wroc$weights,
              tag.nonevent = 0,
              cutoff.value = 0.5)
sp.obj
}
