% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wauc.R
\name{wauc}
\alias{wauc}
\title{Estimation of the AUC of logistic regression models with complex survey data.}
\usage{
wauc(
  response.var,
  phat.var,
  weights.var = NULL,
  tag.event = NULL,
  tag.nonevent = NULL,
  data = NULL,
  design = NULL
)
}
\arguments{
\item{response.var}{Vector with information of the response variable
or character string with the name of the variable
indicating the response variable in the data set.}

\item{phat.var}{Vector with information of the estimated probabilities
or character string with the name of the variable
indicating these probabilities in the data set.}

\item{weights.var}{Vector with information of the sampling weights
or character string with the name of the variable
indicating the sampling weights in the data set.
If \code{NULL}, the sampling design should be specified in the argument \code{design}.
For unweighted estimates, set all the sampling weight values to 1.}

\item{tag.event}{Label indicating the event of interest in \code{response.var}.
The default option is \code{tag.event = NULL}, which selects the class with the lowest number of units.}

\item{tag.nonevent}{Label indicating the non-event in \code{response.var}.
The default option is \code{tag.nonevent = NULL}, which selects the class with the greatest number of units.}

\item{data}{Data frame which must incorporate information on the columns
\code{response.var}, \code{phat.var} and \code{weights.var}.
If \code{data=NULL}, then specific numerical vectors must be included in
\code{response.var}, \code{phat.var} and \code{weights.var}
or the sampling design should be indicated in the argument \code{design}.}

\item{design}{An object of class \code{survey.design} generated by
\code{survey::svydesign} indicating the complex sampling design of the data.
If \code{NULL} information on the data and weights must be included in the argument
\code{data} or as a vector in the argument \code{weights.var}.}
}
\value{
The output object of this function is a list of 4 elements containing the following information:
\itemize{
\item \code{AUCw}: the weighted estimate of the AUC.
\item \code{tags}: a list containing two elements with the following information:
\itemize{
\item \code{tag.event}: a character string indicating the event of interest.
\item \code{tag.nonevent}: a character string indicating the non-event.
}
\item \code{basics}: a list containing information of the following 4 elements:
\itemize{
\item \code{n.event}: number of units with the event of interest in the data set.
\item \code{n.nonevent}: number of units without the event of interest in the data set.
\item \code{hatN.event}: number of units with the event of interest represented in the population by all the event units in the data set, i.e., the sum of the sampling weights of the units with the event of interest in the data set.
\item \code{hatN.nonevent}: a numeric value indicating the number of non-event units in the population represented by means of the non-event units in the data set, i.e., the sum of the sampling weights of the non-event units in the data set.
}
\item \code{call}: an object saving the information about the way in which the function has been run.
}
}
\description{
Calculate the AUC of a logistic regression model
considering sampling weights with complex survey data
}
\examples{
data(example_data_wroc)

auc.obj <- wauc(response.var = "y",
                phat.var = "phat",
                weights.var = "weights",
                tag.event = 1,
                tag.nonevent = 0,
                data = example_data_wroc)

# Or equivalently
auc.obj <- wauc(response.var = example_data_wroc$y,
                phat.var = example_data_wroc$phat,
                weights.var = example_data_wroc$weights,
                tag.event = 1, tag.nonevent = 0)

}
