% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{wroc}
\alias{wroc}
\title{Estimation of the ROC curve of logistic regression models with complex survey data}
\usage{
wroc(response.var, phat.var, weights.var, event = 1, nonevent = 0, data = NULL)
}
\arguments{
\item{response.var}{vector with information of the response variable
or character string indicating the name of the variable
indicating the response variable in the data set}

\item{phat.var}{vector with information of the estimated probabilities
or character string indicating the name of the variable
indicating these probabilities in the data set}

\item{weights.var}{vector with information of the sampling weights
or character string indicating the name of the variable
indicating the sampling weights in the data set}

\item{event}{label indicating the event of interest (default \code{event = 1})}

\item{nonevent}{label indicating the non-event (default \code{nonevent = 0})}

\item{data}{name of the data set.
If data=NULL, then specific numerical vectors must be included in
\code{response.var}, \code{phat.var} and \code{weights.var}}
}
\value{
A graphic with the ROC curve
}
\description{
Calculate the ROC curve of a logistic regression model
considering sampling weights with complex survey data
}
\examples{
data(example_data_wroc)

wroc(response.var = "y", phat.var = "phat", weights.var = "weights", data = example_data_wroc)

# Or equivalently
wroc(response.var = example_data_wroc$y,
     phat.var = example_data_wroc$phat,
     weights.var = example_data_wroc$weights)

}
